# golang/resource/v1/resource.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "resource.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Resource"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/inner/v1/resource/attachment/create": {
      "post": {
        "summary": "创建附件",
        "operationId": "Resource_CreateAttachment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAttachmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAttachmentRequest"
            }
          }
        ],
        "tags": [
          "Resource"
        ]
      }
    },
    "/inner/v1/resource/attachment/delete": {
      "post": {
        "summary": "移除附件",
        "operationId": "Resource_DeleteAttachment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteAttachmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DeleteAttachmentRequest"
            }
          }
        ],
        "tags": [
          "Resource"
        ]
      }
    },
    "/inner/v1/resource/attachment/list": {
      "post": {
        "summary": "附件列表",
        "operationId": "Resource_ListAttachments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAttachmentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListAttachmentsRequest"
            }
          }
        ],
        "tags": [
          "Resource"
        ]
      }
    },
    "/inner/v1/resource/file/list": {
      "post": {
        "summary": "文件列表",
        "operationId": "Resource_ListFiles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListFilesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListFilesRequest"
            }
          }
        ],
        "tags": [
          "Resource"
        ]
      }
    },
    "/inner/v1/resource/getPolicy": {
      "post": {
        "operationId": "Resource_GetPolicy",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPolicyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetPolicyRequest"
            }
          }
        ],
        "tags": [
          "Resource"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateAttachmentRequest": {
      "type": "object"
    },
    "v1CreateAttachmentResponse": {
      "type": "object"
    },
    "v1DeleteAttachmentRequest": {
      "type": "object"
    },
    "v1DeleteAttachmentResponse": {
      "type": "object"
    },
    "v1GetPolicyRequest": {
      "type": "object"
    },
    "v1GetPolicyResponse": {
      "type": "object"
    },
    "v1ListAttachmentsRequest": {
      "type": "object"
    },
    "v1ListAttachmentsResponse": {
      "type": "object"
    },
    "v1ListFilesRequest": {
      "type": "object"
    },
    "v1ListFilesResponse": {
      "type": "object"
    }
  }
}